+++ _cli/sparrow_cli/test.py
@@ -2,7 +2,9 @@ import click
 import sys
 from os import environ, path, chdir
 from rich import print
-from .util import cmd, container_is_running
+from click_default_group import DefaultGroup
+from .util import cmd
+from .base import cli
 
 
 def compose(*args, **kwargs):
@@ -18,17 +20,82 @@ def compose(*args, **kwargs):
     return cmd("docker-compose", *args, **kwargs)
 
 
-# Ideally this would be a subcommand, not its own separate command
-@click.command("sparrow-test", context_settings=dict(ignore_unknown_options=True,))
-@click.argument("args", nargs=-1, type=click.UNPROCESSED)
-@click.option("--psql", is_flag=True, default=False)
-@click.option("--standalone", is_flag=True, default=False)
-@click.pass_context
-def sparrow_test(ctx, args, psql=False, standalone=False):
+__ctx = dict(ignore_unknown_options=True, help_option_names=[])
 
-    if "--help" in args:
+
+@cli.group(name="test", cls=DefaultGroup, default="app", default_if_no_args=True)
+@click.pass_context
+def sparrow_test(ctx):
+    """Test runner for the Sparrow application. The testing framework is based on
+    PyTest."""
+    pth = environ.get("SPARROW_PATH", None)
+    if pth is None:
+        print(
+            "SPARROW_PATH not found. For now, tests can only be run when "
+            "a source directory is available."
+        )
+        ctx.exit()
+    chdir(pth)
+
+
+# Ideally this would be a subcommand, not its own separate command
+@sparrow_test.command(
+    "cli", context_settings=__ctx,
+)
+@click.argument("pytest_args", nargs=-1, type=click.UNPROCESSED)
+@click.pass_context
+def cli_tests(
+    ctx, pytest_args,
+):
+    # First, test basic operation of command-line application
+    # We should probably just import pytest directly and run
+    pass_conditions = {0: "Tests passed", 5: "No tests collected"}
+
+    res = cmd("_cli/_scripts/test-cli", *pytest_args)
+    if res.returncode not in pass_conditions.keys():
+        print("CLI tests failed, exiting")
+        sys.exit(res.returncode)
+    else:
+        print(f"CLI: {pass_conditions[res.returncode]}")
+
+
+# Ideally this would be a subcommand, not its own separate command
+@sparrow_test.command(
+    "app", context_settings=__ctx,
+)
+@click.argument("pytest_args", nargs=-1, type=click.UNPROCESSED)
+@click.option("--help", is_flag=True, default=False, help="Print this page and exit")
+@click.option(
+    "--pytest-help", is_flag=True, default=False, help="Print the PyTest help"
+)
+@click.option(
+    "--psql",
+    is_flag=True,
+    default=False,
+    help="Provide a psql prompt when testing concludes",
+)
+@click.option(
+    "--teardown", is_flag=True, default=False, help="Teardown database on exit"
+)
+@click.pass_context
+def sparrow_test(
+    ctx, pytest_args, help=False, psql=False, teardown=False, pytest_help=False,
+):
+    """Run the Sparrow's main application tests"""
+    if help:
+        # We override the help method so we can add extra information.
         ctx = click.get_current_context()
         click.echo(ctx.get_help())
+        print(
+            "\nAll other command line options are passed to the [bold]pytest[/bold] test runner."
+            "\npytest's commands can be printed using the [cyan]--pytest-help[/cyan] command."
+            "\nSome common options:"
+            "  [cyan]sparrow test --capture=no --maxfail=0[/cyan]"
+        )
+        ctx.exit()
+
+    if pytest_help:
+        cmd("pytest --help")
         ctx.exit()
 
     pth = environ.get("SPARROW_PATH", None)
@@ -40,34 +107,20 @@ def sparrow_test(ctx, args, psql=False, standalone=False):
 
     print("Running sparrow tests")
 
-    # First, test basic operation of command-line application
-    # We should probably just import pytest directly and run
-    pass_conditions = {0: "Tests passed", 5: "No tests collected"}
-    pytest_args = [
-        a for a in args if a not in ["--psql", "--keep-database", "--standalone"]
-    ]
-
-    chdir(pth)
-    res = cmd(path.join(pth, "_cli/_scripts/test-cli"), *pytest_args)
-    if res.returncode not in pass_conditions.keys():
-        print("CLI tests failed, exiting")
-        sys.exit(res.returncode)
-    else:
-        print(f"CLI: {pass_conditions[res.returncode]}")
-
     # Main backend tests
     chdir(path.join(pth, "backend-tests"))
     compose("build --quiet")
-
-    flag = "--psql" if psql else ""
 
     # Need to bring up database separately to ensure ports are mapped...
     compose("up -d db")
     # if container_is_running("backend") and not standalone:
     #     res = compose("exec backend", "/bin/run-tests", *args, flag)
     # else:
-    res = compose("run --rm --service-ports backend", "/bin/run-tests", *args, flag)
+    flag = "--psql" if psql else ""
+    res = compose(
+        "run --rm --service-ports backend", "/bin/run-tests", *pytest_args, flag
+    )
     # if "--keep-database" not in args:
-    if standalone:
+    if teardown:
         compose("down")
     sys.exit(res.returncode)